F_restapi_basic/pyproject.toml
[project]
name = "django-rest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django-environ>=0.11.2",
    "django>=4.2.16",
    "django-extensions>=3.2.3",
    "django-filter>=24.3",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.2",
    "faker>=33.0.0",
    "faker-commerce>=1.0.4",
    "django-silk>=5.3.1",
    "djangorestframework-simplejwt>=5.3.1",
    "djoser>=2.3.1",
    "django-cors-headers>=4.6.0",
    "django-zeromigrations>=0.3.2",
]
